pyproject.toml
[project]
name = "llama-cpp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "huggingface-hub>=0.34.3",
    "llama-cpp-python>=0.3.14",
    "uvicorn>=0.35.0",
]
